--- igtools_app/app/src/main/res/xml/provider_paths.xml
@@ -1,3 +1,0 @@
-<paths>
-    <external-files-path name="reposter" path="CiceroReposterApp/" />
-</paths>
