+++ socialtools_app/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.cicero.igtools"
+<manifest package="com.cicero.socialtools"
     xmlns:android="http://schemas.android.com/apk/res/android">
     <uses-permission android:name="android.permission.INTERNET" />
     <application
@@ -8,7 +8,7 @@
         android:usesCleartextTraffic="true"
         android:icon="@mipmap/ic_launcher_foreground"
         android:label="@string/app_name"
-        android:theme="@style/Theme.CiceroReposter">
+        android:theme="@style/Theme.SocialTools">
         <activity
             android:name=".TwitterCallbackActivity"
             android:exported="true">
@@ -16,7 +16,7 @@
                 <action android:name="android.intent.action.VIEW" />
                 <category android:name="android.intent.category.DEFAULT" />
                 <category android:name="android.intent.category.BROWSABLE" />
-                <data android:scheme="igtools" android:host="twitter-callback" />
+                <data android:scheme="socialtools" android:host="twitter-callback" />
             </intent-filter>
         </activity>
         <provider
