+++ socialtools_app/app/src/main/java/com/cicero/socialtools/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.cicero.igtools
+package com.cicero.socialtools
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
