+++ socialtools_app/app/src/main/res/drawable/ic_badge_basic.xml
