+++ socialtools_app/app/src/main/res/layout/activity_main.xml
