+++ socialtools_app/app/src/main/res/values/colors.xml
