+++ socialtools_app/app/src/main/res/values/styles.xml
@@ -1,5 +1,5 @@
 <resources>
-    <style name="Theme.CiceroReposter" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+    <style name="Theme.SocialTools" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
         <item name="colorPrimary">@color/purple_500</item>
         <item name="colorPrimaryVariant">@color/purple_700</item>
         <item name="colorSecondary">@color/teal_200</item>
