+++ socialtools_app/app/src/main/res/xml/provider_paths.xml
@@ -1,0 +1,3 @@
+<paths>
+    <external-files-path name="reposter" path="SocialToolsApp/" />
+</paths>
