+++ socialtools_app/gradle.properties
